refactor preprocess to different layers
